GPP - wip(rotations): Rotation supports
Rotation supports can have inner rotation and rotate player if he's on it (thanks Unreal physics)
Ctrl+R now checks for a Rotation Support that the player is on and start a debug rotation on it.
WIP of rotating objects that are on Rotation Support.

BP_RotationSupport
BP_SceneRotationManager
BP_SceneManager
BPAC_SceneRotationObject

LVL_Scene_1