GPP - wip(movable): movable objects merge
Merging objects that can move through rotation support and moving platform so that everything can work together.
Currently only debug movements and rotations are functional, "static" movable objects are working, but there is issues with objects that inherit a movement of the same type that they create (inherited rotation for rotation support for exemple).
Also cleaned a bit the file organization.

Every blueprint that is linked to rotations or moving platform (no BPPC_Player)