GPP - wip(control): Moving objects control
Added a secondary interaction that takes the most suitable moving object according to player position and either apply a Movement to it (moving platform) or select it for control (rotation support). Also added an input mode allowing to control the selected rotation support.
Code is work in progress and features of this controlled rotations are missing. (like snap and clamp)